Modified driver test functions for general usage
Modified the testingOutJDBC function to have String queries that will work regardless of where the database is hosted.

Convert the string queries to present create their own table and access only their
own table so that even if
the query is executed on different machines
it will work all the same.