Created Classes for DB tables with Creation String
Created Package for testing the program. Moved App.java inside
the package, the App.java needs to be renamed for the specified
purpose.

Created Java Classes to represent the data of a table inside the
database. The Classes will implement all the CRUD operation for the
represented table.

The class will contain their own specialized String of MySQL queries.