Added Class & Empty Method Body for db initializer
Created a class whose purpose is to communicate with the database
and execute query inside of the database.

As of now, the bodies of the methods are empty because I don't have a
concrete idea on how the query should be handled.

Very small changes inside the db table classes, removed and added some
spaces, for readability purposes.

The body of the Menuitems class is empty because I will first
prioritize the creation of the class responsible for connecting and
initializing the project database.