Start progress: creating proj db & support methods
1. MAJOR CHANGES
Started working on the progress for the initialization class,
I am working on the finalization of the process for creating
the database inside the server first.

The method for creating the database is unfinished, I shifted my focus
on creating a method to identify if a database already exist inside the
server.

I don't know how to create a MySQL query to create a Database\Table
if the Database\Table does not exist yet.

2. Minor changes

Created a package to hold the classes to handle the database tables.
Moved the empty Menuitems class.