Working code to create table inside proj db.
Created a method to create table inside the project database using an
instance of the DBEntity interface.

Transferred the funtionality for creating the database table inside the
DatabaseInitialize class from the DBEntity interface to reduce
repititive code.

Renamed App.java to DatabaseConnectionReference.java because the original
name is too vague to explain it's uses.