Added Msgs to inform for connection server problem
Did not print the stacktrace, instead I added a JOptionPane Message
dialog to present a simplified explanation of the problem.

It is possible to show the stacktrace by asking the user if they want
the stack trace.

The functions that calls for a Connection Object with the server must
handle the null return value.

I think I should comment it out or add a throws declaration in the
method signature.