Fixed Merging inconsistencies of Main and branch
Merging the main and the branch for the databaseInitialization
resulted into errors because the names of packages were changed.

The DBEntity interface caused errors because there are classes that
implemented the old version of the class, the version with the
deprecated functions (removed functions) and missing new functions.

Had to update those classes, remove some methods and added
implementations of the new method.