Added /.vscode inside gitignore
Added the /.vscode folder inside the gitignore file because the folder
involves the configuration of the workspace of the device that is
working on a clone of the repository.

Which means for every device, the chance that the configuration
is unique is very high.

To avoid merge conflicts and configuration problems in the future.

Configuration folder and files should not be tracked and added to
the repository.

Fixed a minor import statement inside databaseDevelopment.java.