Modified Table Classes to SPPRT DBEntity interface
Added a variable to represent the name of the table.

Embedded the variable name inside the table creation MySQL query.

The table names are all small letters because inside the MySQL
database, when checking the database if a table exists inside, the
table name variable can be used for finding the table name inside the
project database.