Working Feature to Create all tables inside DB
NOTE: Will not be specifying the method and class names because
they are subject to change during development. The ideas and intent
can be considered constant, but the references/name is guaranteed to
have changes.

NOTE: Method to Create All Tables.

Param for creataAllProjectTables(): An Array of DBEntity

Created an Array to store the references of the instances of the
DB table entities/classes.

There are two types of DBEntity, the distinction between
the Strong and Weak entities are important because the Strong entity
needs to be created first before any of the weak entities are created.

In a scenario when the Weak Entity is created first before any of the
referenced table, this throws an exception.

Table Creation Function:
Uses a For-each loop to iterate over the array of DBEntity, for each
DBEntity, the DBEntity is passed over the method to Create a table
inside the project database.

Support Function:
Created a function to check if the table already exist inside the
project database. The function returns a boolean variable of true if
the table exists inside the project database.

If the result is true, the table creation function will not proceed with
executing the MySQL query to create the table.