Added method comments for Creating all tables
The parameter for the method that creates all tables is an
array of DBEntity(interface).

Added a comment to explain the details regarding the parameter
and the method that creates the tables using the parameter.

Interfaces don't have method body but method names only. Allowing
for different Classes or Objects to be identified and classified
as the interface they are implementing.

This is useful for the situation wherein the Classes that
represents a table entity will have huge differences in terms of
capabilities and behaviour.

There is a need to identify them as one
type and access certain methods.

Implementing an interface guarantees both of the conditions will be
fulfilled.