Removed Menuitems, modified other DBEntities
Changed the MySQL query creation string Employees to allow for
password columns and authority flags(boolean variable columns).

Added some minor explanatory comments.

Removed the Menuitems class because the source control tab for changes
always keeping on showing it, I tried to add it to the gitignore- but
the changes tab still shows it.

Removed it so that it won't annoy me anymore. The use case for the
DBEntity is none, since project will shift towards Customer
Reservation Management.