WORKING Registration And Removal\Flagging Feature
Flagging - Change the boolean column/variable that represents the
authority level and working status of an employee inside the database.

Registration Process for Employees is working, both for Admin and
Non-Admin Employees.

Removing an Entry inside the Employee table is working. The
Entry to be removed is identified by the employeeid.

Major modifications to the method to register admin users.

1.Dismantled the method to register admin users.

2.Greatly reduced method size through delegating the code inside to
separate methods.

CHANGES:

Created separate methods for producing the MySQL queries.

The MySQL queries are resusable , there are portions of the program
that may need the queries.

The Registration process for Employees and
User alike are very similar. This means that the code will most likely
be the same except for some minor changes.

This would mean that the code for both methods will mostly likely be
the same size, same words, same results, aside for different
values in the authority flag ( MySQL column for determining if Employee
is an admin or not. A boolean variable that determines if an Employee
is an admin or not ).

REMOVING AN EMPLOYEE AND FLAGGING AN EMPLOYEE AS NON-WORKING

Method for Removing an Entry inside the Employee Table is untested and
yet to be checked for exceptions and problems. The method uses the
same WHERE clause that the UPDATE statements used inside the Program.

Flagging an employee as Non-working turns their Working and authority
variable/column into false, since the Working and authority Status is
represented by a boolean variable.

REASON FOR CHANGES:

To prevent code duplication and overly long code.

I used method to contain portions of the code that can be called
when registering a user(admin or employee).

The method for the registration of the Employee and the Admin will
use mostly the same method calls with slight differences.

I used multiple Update statements over a single Insert statement
because I want to identify which portion in user input there is
a problem with.

After identifying which portion did the user mishandled
the process, proper feed back can be formulated to the user.