WORKING READ Operations on Employees
Working :
  Log-in verification
  Employee Column Data Retrieval
  Employee Row Data Retrieval through Employee ID Identification

Added comments to Most Methods inside the Employee Class, Explanation
of the Method, though the use cases and intent behind the method needs
to be explained in the comments in a way that can be read and
understood as fast as possible.

Created methods to test the Read Methods for the Employees table.
Messy modifications in DatabaseDevelopmentFeaturesTester, no comments
and/or vague comments.

Misplaced or misused Print statements. Will have to clean up when
program undergoes demonstration.