Added Empty Methods & Constructor For Bridge Class
Created A Bridge Class, A class/object to bridge the user and the
database.

The Class will handle the interaction between the user and program,
by extension, the class will translate the actions of the user
to C.R.U.D actions for the tables inside the Database.

I created Empty Methods first because I don't know what code to write
to provide an interaction between user and the program that involves
data from the incomplete database.

The empty methods helps in providing me an idea for future progress.