Created User Authority Identification
Created Functions for Supporting identification process.
Modified overall structure of the ProgramUser Constructor to execute
one of two methods for CLI or GUI mode.

Currently for testing and Development, the CLI methods and Program
Runtime is prioritized for development.

Added a Helper Utility Class to handle printing the ResultSet from the
Tables, which will be important for the Read Functions in the CLI
version of the program.