Created new Additions for Table Classes
Created the Rooms and CustomerService table classes to meet the
requirements for 5 tables. The C.R.U.D operations of these classes
will be conducted through calls to the Utility class for Tables.

Made some minor changes to the Table Classes, each of which will
have variables to represent their own columns and table name. If these
variables are String then they are all lower case letters because
MySQL presents String characters in Lower case, this makes it
easier for making WHERE statements.