Working LogIn Form For Admin Employee
Created package to contain images used for GUI Components.

Created code inside the ProgramUser constructor for the GUI version
of the program. Added code inside the LogInGUIVersion.

Created a JFrame to display the Data of a table. To be tested for
stability and compatibility with other tables, tested with Reservations
table.