WORKING VIEW Table Functionality & Code Reformat
Reformatted the code inside the GUI for Adding Reservation,
Adding Employees, and the code for AdminDashboard for convenience and
 familiarity over the structure of the program.

Needed to label the variables, for finding the JButton variables, I
really need to know where the reference to these JButtons because I
need to attach the actionlistener to them.

Working functionality to view the contents fo the database tables
through an external jframe with a jtable.